namespace all outbound ECF API service classes
