+++ app/lib/services/ecf/ecf_user_finder.rb
@@ -1,0 +1,15 @@
+module Services
+  module Ecf
+    class EcfUserFinder
+      attr_reader :user
+
+      def initialize(user:)
+        @user = user
+      end
+
+      def call
+        EcfApi::User.where(email: user.email).first
+      end
+    end
+  end
+end
