--- app/lib/services/ecf_user_creator.rb
@@ -1,15 +1,0 @@
-module Services
-  class EcfUserCreator
-    attr_reader :user
-
-    def initialize(user:)
-      @user = user
-    end
-
-    def call
-      remote = EcfApi::User.new(email: user.email, full_name: user.full_name)
-      remote.save
-      user.update!(ecf_id: remote.id)
-    end
-  end
-end
