--- app/lib/services/ecf_user_finder.rb
@@ -1,13 +1,0 @@
-module Services
-  class EcfUserFinder
-    attr_reader :user
-
-    def initialize(user:)
-      @user = user
-    end
-
-    def call
-      EcfApi::User.where(email: user.email).first
-    end
-  end
-end
