--- app/lib/services/npq_profile_updater.rb
@@ -1,15 +1,0 @@
-module Services
-  class NpqProfileUpdater
-    attr_reader :application
-
-    def initialize(application:)
-      @application = application
-    end
-
-    def call
-      profile = EcfApi::NpqProfile.find(application.ecf_id).first
-      profile.eligible_for_funding = application.eligible_for_funding
-      profile.save
-    end
-  end
-end
