+++ lib/tasks/sync_applications.rake
@@ -8,7 +8,7 @@ namespace :sync do
     Application.where("ecf_id is not null").order(created_at: :asc).each_with_index do |application, i|
       sleep(0.1)
       Rails.logger.info "syncing application #{application.id}, (#{i + 1},/#{count})"
-      Services::NpqProfileUpdater.new(application: application).call
+      Services::Ecf::NpqProfileUpdater.new(application: application).call
     rescue StandardError
       Rails.logger.info "error with application #{application.id}"
       errored_ids << application.id
