+++ spec/jobs/application_submission_job_spec.rb
@@ -13,13 +13,13 @@ RSpec.describe ApplicationSubmissionJob do
     end
 
     it "calls correct services" do
-      user_finder_double = instance_double(Services::EcfUserFinder, call: nil)
-      user_creator_double = instance_double(Services::EcfUserCreator, call: nil)
-      profile_creator_double = instance_double(Services::NpqProfileCreator, call: nil)
+      user_finder_double = instance_double(Services::Ecf::EcfUserFinder, call: nil)
+      user_creator_double = instance_double(Services::Ecf::EcfUserCreator, call: nil)
+      profile_creator_double = instance_double(Services::Ecf::NpqProfileCreator, call: nil)
 
-      expect(Services::EcfUserFinder).to receive(:new).with(user: user).and_return(user_finder_double)
-      expect(Services::EcfUserCreator).to receive(:new).with(user: user).and_return(user_creator_double)
-      expect(Services::NpqProfileCreator).to receive(:new).with(application: application).and_return(profile_creator_double)
+      expect(Services::Ecf::EcfUserFinder).to receive(:new).with(user: user).and_return(user_finder_double)
+      expect(Services::Ecf::EcfUserCreator).to receive(:new).with(user: user).and_return(user_creator_double)
+      expect(Services::Ecf::NpqProfileCreator).to receive(:new).with(application: application).and_return(profile_creator_double)
 
       subject.perform_now
 
@@ -28,13 +28,13 @@ RSpec.describe ApplicationSubmissionJob do
     end
 
     it "sends submission email" do
-      user_finder_double = instance_double(Services::EcfUserFinder, call: nil)
-      user_creator_double = instance_double(Services::EcfUserCreator, call: nil)
-      profile_creator_double = instance_double(Services::NpqProfileCreator, call: nil)
+      user_finder_double = instance_double(Services::Ecf::EcfUserFinder, call: nil)
+      user_creator_double = instance_double(Services::Ecf::EcfUserCreator, call: nil)
+      profile_creator_double = instance_double(Services::Ecf::NpqProfileCreator, call: nil)
 
-      expect(Services::EcfUserFinder).to receive(:new).with(user: user).and_return(user_finder_double)
-      expect(Services::EcfUserCreator).to receive(:new).and_return(user_creator_double)
-      expect(Services::NpqProfileCreator).to receive(:new).and_return(profile_creator_double)
+      expect(Services::Ecf::EcfUserFinder).to receive(:new).with(user: user).and_return(user_finder_double)
+      expect(Services::Ecf::EcfUserCreator).to receive(:new).and_return(user_creator_double)
+      expect(Services::Ecf::NpqProfileCreator).to receive(:new).and_return(profile_creator_double)
 
       allow(ApplicationSubmissionMailer).to receive(:application_submitted_mail).and_call_original
       expect(ApplicationSubmissionMailer).to receive(:application_submitted_mail).with(
@@ -51,11 +51,11 @@ RSpec.describe ApplicationSubmissionJob do
       let(:user) { create(:user, ecf_id: "123") }
 
       it "calls correct servivces" do
-        instance_double(Services::EcfUserCreator)
-        profile_creator_double = instance_double(Services::NpqProfileCreator, call: nil)
+        instance_double(Services::Ecf::EcfUserCreator)
+        profile_creator_double = instance_double(Services::Ecf::NpqProfileCreator, call: nil)
 
-        expect(Services::EcfUserCreator).not_to receive(:new)
-        expect(Services::NpqProfileCreator).to receive(:new).with(application: application).and_return(profile_creator_double)
+        expect(Services::Ecf::EcfUserCreator).not_to receive(:new)
+        expect(Services::Ecf::NpqProfileCreator).to receive(:new).with(application: application).and_return(profile_creator_double)
 
         subject.perform_now
 
@@ -68,12 +68,12 @@ RSpec.describe ApplicationSubmissionJob do
       let(:ecf_user) { EcfApi::User.new(email: user.email, id: "123") }
 
       it "calls correct servivces" do
-        user_finder_double = instance_double(Services::EcfUserFinder, call: ecf_user)
-        profile_creator_double = instance_double(Services::NpqProfileCreator, call: nil)
+        user_finder_double = instance_double(Services::Ecf::EcfUserFinder, call: ecf_user)
+        profile_creator_double = instance_double(Services::Ecf::NpqProfileCreator, call: nil)
 
-        expect(Services::EcfUserFinder).to receive(:new).with(user: user).and_return(user_finder_double)
-        expect(Services::EcfUserCreator).not_to receive(:new)
-        expect(Services::NpqProfileCreator).to receive(:new).with(application: application).and_return(profile_creator_double)
+        expect(Services::Ecf::EcfUserFinder).to receive(:new).with(user: user).and_return(user_finder_double)
+        expect(Services::Ecf::EcfUserCreator).not_to receive(:new)
+        expect(Services::Ecf::NpqProfileCreator).to receive(:new).with(application: application).and_return(profile_creator_double)
 
         subject.perform_now
 
@@ -89,11 +89,11 @@ RSpec.describe ApplicationSubmissionJob do
       let(:application) { create(:application, user: user, ecf_id: "456", school: school) }
 
       it "calls correct servivces" do
-        instance_double(Services::EcfUserCreator)
-        instance_double(Services::NpqProfileCreator, call: nil)
+        instance_double(Services::Ecf::EcfUserCreator)
+        instance_double(Services::Ecf::NpqProfileCreator, call: nil)
 
-        expect(Services::EcfUserCreator).not_to receive(:new)
-        expect(Services::NpqProfileCreator).not_to receive(:new)
+        expect(Services::Ecf::EcfUserCreator).not_to receive(:new)
+        expect(Services::Ecf::NpqProfileCreator).not_to receive(:new)
 
         subject.perform_now
       end
